% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifa.R
\name{compute_log_likelihood}
\alias{compute_log_likelihood}
\title{Compute log-likelihood for current parameter estimates}
\usage{
compute_log_likelihood(Y, Z, lambda, sigma, decay_coef)
}
\arguments{
\item{Y}{Data matrix}

\item{Z}{Mean of latent variables}

\item{lambda}{Loading matrix}

\item{sigma}{Vector of variances}

\item{decay_coef}{Zero-inflation decay coefficient}
}
\value{
Log-likelihood value
}
\description{
Compute log-likelihood for current parameter estimates
}
\keyword{internal}
