% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifa.R
\name{compute_posterior_Z}
\alias{compute_posterior_Z}
\title{Compute posterior distribution of Z given Y}
\usage{
compute_posterior_Z(Y, lambda, sigma, decay_coef)
}
\arguments{
\item{Y}{Data matrix}

\item{lambda}{Loading matrix}

\item{sigma}{Vector of variances}

\item{decay_coef}{Zero-inflation decay coefficient}
}
\value{
List with posterior mean and covariance
}
\description{
Compute posterior distribution of Z given Y
}
\keyword{internal}
