% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifa.R
\name{decay_log_likelihood}
\alias{decay_log_likelihood}
\title{Log-likelihood for the decay coefficient}
\usage{
decay_log_likelihood(Y, Z, lambda, sigma, decay_coef)
}
\arguments{
\item{Y}{Data matrix}

\item{Z}{Mean of latent variables}

\item{lambda}{Loading matrix}

\item{sigma}{Vector of variances}

\item{decay_coef}{Zero-inflation decay coefficient}
}
\value{
Log-likelihood value for decay coefficient
}
\description{
Log-likelihood for the decay coefficient
}
\keyword{internal}
