% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifa.R
\name{initialize_params}
\alias{initialize_params}
\title{Initialize ZIFA parameters}
\usage{
initialize_params(Y, K, single_sigma = FALSE)
}
\arguments{
\item{Y}{Data matrix (N x D)}

\item{K}{Number of latent dimensions}

\item{single_sigma}{Use single variance}
}
\value{
List of initial parameters
}
\description{
Initialize ZIFA parameters
}
\keyword{internal}
