% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifa.R
\name{Mstep}
\alias{Mstep}
\title{M-step: Update model parameters}
\usage{
Mstep(Y, EZ, EZZT, EX, EX2, old_decay_coef, single_sigma = FALSE)
}
\arguments{
\item{Y}{Data matrix (N x D)}

\item{EZ}{Expected Z (N x K)}

\item{EZZT}{Expected ZZ' (N x K x K)}

\item{EX}{Expected X (N x D)}

\item{EX2}{Expected X^2 (N x D)}

\item{old_decay_coef}{Previous decay coefficient}

\item{single_sigma}{If TRUE, use single variance}
}
\value{
Updated parameters
}
\description{
M-step: Update model parameters
}
\keyword{internal}
