% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_zifa}
\alias{plot_zifa}
\title{Plot ZIFA Results}
\usage{
plot_zifa(Z, labels = NULL, dims = c(1, 2))
}
\arguments{
\item{Z}{Matrix of ZIFA projections}

\item{labels}{Optional vector of labels for coloring points}

\item{dims}{Dimensions to plot (default: first two dimensions)}
}
\value{
A ggplot2 object
}
\description{
Creates a scatter plot of ZIFA results in two dimensions.
}
\examples{
# Generate simulated data
set.seed(123)
Y <- matrix(rpois(1000, lambda = 5), nrow = 100)
Y[sample(length(Y), 500)] <- 0
result <- fit_zifa(Y, k = 2)
# Plot results
if (requireNamespace("ggplot2", quietly = TRUE)) {
  plot_zifa(result$Z)
}

}
