% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifa.R
\name{update_parameters}
\alias{update_parameters}
\title{Update model parameters}
\usage{
update_parameters(Y, Z_mean, Z_cov, decay_coef)
}
\arguments{
\item{Y}{Data matrix}

\item{Z_mean}{Posterior means of Z}

\item{Z_cov}{Posterior covariances of Z}

\item{decay_coef}{Current decay coefficient}
}
\value{
List of updated parameters
}
\description{
Update model parameters
}
\keyword{internal}
