% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifa.R
\name{validate_input}
\alias{validate_input}
\title{Validate input data for ZIFA}
\usage{
validate_input(Y)
}
\arguments{
\item{Y}{Data matrix (samples x genes)}
}
\value{
NULL invisibly if valid, otherwise throws error
}
\description{
Validate input data for ZIFA
}
\keyword{internal}
