% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifa.R
\docType{package}
\name{zifa}
\alias{ZIFAR}
\alias{ZIFAR-package}
\alias{zifa}
\title{Zero-Inflated Factor Analysis (ZIFA)}
\description{
Native R implementation of Zero-Inflated Factor Analysis for
dimensionality reduction of single-cell gene expression data.
Based on Pierson & Yau (2015) \url{doi:10.1186/s13059-015-0805-z}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mr-september/ZIFAR}
  \item Report bugs at \url{https://github.com/mr-september/ZIFAR/issues}
}

}
\author{
\strong{Maintainer}: Larry Cai \email{larry.cai@uq.edu.au}

Authors:
\itemize{
  \item Emma Pierson [copyright holder]
  \item Christopher Yau [copyright holder]
}

}
\keyword{internal}
