% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifa.R
\name{compute_log_likelihood}
\alias{compute_log_likelihood}
\title{Compute log-likelihood for monitoring convergence}
\usage{
compute_log_likelihood(Y, EZ, A, mus, sigmas, decay_coef)
}
\description{
Compute log-likelihood for monitoring convergence
}
\keyword{internal}
