% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifa.R
\name{compute_zero_probability}
\alias{compute_zero_probability}
\title{Compute Zero Probability}
\usage{
compute_zero_probability(mu, decay_coef)
}
\arguments{
\item{mu}{Mean of the normal component}

\item{decay_coef}{Decay coefficient}
}
\value{
Probability of zero inflation
}
\description{
Compute Zero Probability
}
\keyword{internal}
