% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_zifa.R
\name{fit_block_zifa}
\alias{fit_block_zifa}
\title{Fit Block ZIFA Model}
\usage{
fit_block_zifa(Y, k, n_blocks = NULL, ...)
}
\arguments{
\item{Y}{Matrix of expression data (genes as rows, cells as columns)}

\item{k}{Integer specifying the number of latent dimensions}

\item{n_blocks}{Number of blocks to divide genes into (default: genes/500)}

\item{...}{Additional parameters passed to fit_zifa}
}
\value{
A list containing:
\item{Z}{Matrix of low-dimensional projections (cells x k dimensions)}
\item{model_params}{List of model parameters}
}
\description{
Implements Block Zero-Inflated Factor Analysis for more efficient processing
of large datasets by dividing genes into blocks.
}
\examples{
# Generate large simulated dataset
set.seed(123)
true_Z <- matrix(rnorm(100), ncol = 2)
lambda <- matrix(runif(1000*2, -1, 1), nrow = 1000)
mu <- true_Z \%*\% t(lambda)
Y <- mu + matrix(rnorm(1000*50), nrow = 1000)
# Add zero-inflation
mask <- matrix(rbinom(1000*50, 1, prob = 0.3), nrow = 1000)
Y[mask == 1] <- 0

# Run Block ZIFA
result <- fit_block_zifa(Y, k = 2, n_blocks = 2)

}
