% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_zifa}
\alias{plot_zifa}
\title{Plot ZIFA Results}
\usage{
plot_zifa(Z, labels = NULL, dims = c(1, 2))
}
\arguments{
\item{Z}{Matrix of ZIFA projections (cells x dimensions)}

\item{labels}{Optional vector of labels for coloring points (e.g., cell types)}

\item{dims}{Which dimensions to plot (default: first two dimensions)}
}
\value{
A ggplot2 object
}
\description{
Creates a scatter plot of ZIFA results in two dimensions.
}
\examples{
\dontrun{
# After running ZIFA
result <- fit_zifa(Y, k = 2)

# Basic plot
plot_zifa(result$Z)

# With cell type labels
cell_types <- c(rep("TypeA", 25), rep("TypeB", 25))
plot_zifa(result$Z, labels = cell_types)

# Plot dimensions 2 and 3
result_3d <- fit_zifa(Y, k = 3)
plot_zifa(result_3d$Z, dims = c(2, 3))
}

}
